Update README.md
# lyrics-visualiser
An app that visualises lyrics based on their section/line-based self-similiarity. 
It is inspired by https://github.com/colinmorris/SongSim. 
The self-similiarity is however computed based on the song structure (sections) using the jaccard-distance instead of matching single words.
Further the lyrics are loaded using the genius.com api giving it access to a huge lyrics-database.

# links:
<a href="https://github.com/JLiekenbrock/lyrics-visualiser/tree/main/UML">UML<>